# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/portfolios":{"post":{"operationId":"PortfolioController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePortfolioDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/createPortfolioResponse"}}}}}},"get":{"operationId":"PortfolioController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Portfolio"}}}}}}}},"/portfolios/findByName":{"get":{"operationId":"PortfolioController_findByNameId","parameters":[{"name":"portfolio_name_id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Portfolio"}}}}}}},"/portfolios/{id}":{"patch":{"operationId":"PortfolioController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePortfolioDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Portfolio"}}}}}},"delete":{"operationId":"PortfolioController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/portfolios/{id}/positions":{"post":{"operationId":"PortfolioController_addPortfolioPosition","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddPortfolioPositionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/portfolios/{id}/positions/findByStrategy":{"get":{"operationId":"PortfolioController_findPortfolioPositionsByStrategy","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"strategy","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}}}},"/portfolios/positions/{positionId}":{"delete":{"operationId":"PortfolioController_removePortfolioPosition","parameters":[{"name":"positionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Risk management trading tool based on incremental value at risk","description":"This api give you the ability to manage the risk of your trading systems by allowing adding new trades on your portfolio depending the value at risk threshold you are ok with","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreatePortfolioParams":{"type":"object","properties":{"nbComputePeriods":{"type":"number","description":"The number of periods to compute the correlation and standard deviation of the portfolio var","example":20},"zscore":{"type":"number","description":"The zscore used for compute the portfolio var 1.645 for 90% confidence, 1.96 for 95%, 2.58 for 99%","example":1.645},"timeframe":{"type":"string","description":"The timeframe used to compute correlation and standard deviation of the var","example":"15m"},"nameId":{"type":"string","description":"Give the portfolio a unique explicit name","example":"crypto_15m"}},"required":["nbComputePeriods","zscore","timeframe","nameId"]},"CreatePortfolioConstraints":{"type":"object","properties":{"maxVarInDollar":{"type":"number","description":"The max value at risk in dollar for your portfolio","example":1000},"maxOpenTradeSameSymbolSameDirection":{"type":"number","description":"The maximum of open positions for a pair in the same direction (if 1 then you can open only one BTC/USDT long, the second opportunity with BTC/USDT long will be rejected)","example":1}},"required":["maxVarInDollar","maxOpenTradeSameSymbolSameDirection"]},"CreatePortfolioDto":{"type":"object","properties":{"params":{"$ref":"#/components/schemas/CreatePortfolioParams"},"constraints":{"$ref":"#/components/schemas/CreatePortfolioConstraints"}},"required":["params","constraints"]},"createPortfolioResponse":{"type":"object","properties":{"id":{"type":"number","example":1}},"required":["id"]},"Portfolio":{"type":"object","properties":{}},"UpdatePortfolioDto":{"type":"object","properties":{"maxVarInDollar":{"type":"number","minimum":0,"description":"The max value at risk in dollar you would like to change for your portfolio","example":1000},"maxOpenTradeSameSymbolSameDirection":{"type":"number","minimum":1,"description":"The maximum of open positions for a pair in the same direction you would like to change (if 1 then you can open only one BTC/USDT long, the second opportunity with BTC/USDT long will be rejected)","example":1}}},"AddPortfolioPositionDto":{"type":"object","properties":{"pair":{"type":"string","description":"pair to add","example":"BTC/USDT"},"dollarAmount":{"type":"number","minimum":0,"description":"the position amount in dollar","example":1000},"direction":{"type":"string","description":"the direction of the position","example":"long","enum":["long","short"]},"dataSource":{"enum":["binance_futures","binance_spot"],"type":"string","description":"the data source used to get the candles and compute the var","example":"binance_futures"},"strategy":{"type":"string","description":"the name of the strategy link to position","example":"MaCrossOver"}},"required":["pair","dollarAmount","direction","dataSource"]}}}}